form improvements - use ref for submit, clean up type
